+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 todo
 ----
 
-- remove mp3s from github, replace with previews
-- grab audio from system mic
 - import a generic mp3/stream url
 - remove soundcloud sdk.js dependancy
 - debug on ipad
 - grunt uglify & concat css+js
 - add analytics for clientside js errors
+
+v.1.6.0 (2014-11-15)
+--------------------
+- remove mp3s from github, replace with previews
+- grab audio from system mic
 
 
 v.1.5.0 (2014-10-04)